Clone TableDefinitions to avoid static reuse issues.
When using the MSBuild tasks, the tools are loaded once and used
multiple times. That creates the possibility for any static variables
to be incorrectly reused. In this case the Windows Installer standard
table definitions were not getting reset with each run, leaving
extension tables loaded. Cloning the table definitions will avoid the
problem.