Do not cache extension table definitions or libraries.
Static variables can cache the loaded table definitions and libraries
found in extensions. However, there are side effects in the linker and
binder when using cached data across builds (which is possible when
the tools are loaded by MSBuild once). To avoid the side effects, we'll
remove the caching and find better ways to optimize without side effects.