WIP:4259 - Collapse duplicate private directories.
This commit changes the linker to mark duplicate private directory rows
as redundant instead of throwing duplicate symbol errors. Redundant rows
are not output the the .IDTs thus do not end up in the final output.
This allows duplicate private directory rows to collapse into a single
consistent directory tree.