WIP:4259 - Improved duplicate private directory collapsing algorithm.
The ReportDuplicateResolvedSymbolErrorsCommand was designed with the
assumption that there would normally be relatively few duplicates and
that those duplicates would usually be a "bad thing". However, with
the new support for private directory collapsing, the potential for
lots and lots of duplicates greatly increased. This commit improves the
collapsing algorithm to find redundant directory rows when processing
symbols (when we are touch each row already) and store them in a
separate collection from the conflicting symbols. The result is that
the duplicate symbol reporting command is (again) less likely to see
lots of conflicts.